+++ R/pivot_freqs.R
@@ -6,8 +6,9 @@
 #' Given a grouped frequencies table, pivot_freqs will create new columns for each label level in the frequencies
 #'
 #' @param dataset A grouped frequencies table as produced by y2clerk::freqs()
-#'
-#' @return A wide tibble of frequencies with one row for each group
+#' @param columns_var DEFAULT = label; If label, the frequencies will be pivoted so a new column will be created for each unique level of label.
+#' Can also be set to group_var to pivot the other way and create new columns for each unique level of group_var
+#' @return A wide tibble of frequencies with one row for each group (by default)
 #' @export
 #' @examples
 #'   frequencies <- ToothGrowth %>%
@@ -16,8 +17,12 @@
 #'       pivot_freqs()
 
 pivot_freqs <-
-  function(dataset) {
+  function(
+    dataset,
+    columns_var = label
+    ) {
 
+    label <- NULL
     pivot_errors(dataset)
 
     dataset %>%
@@ -27,7 +32,7 @@ pivot_freqs <-
         .data$result
         ) %>%
       tidyr::pivot_wider(
-        names_from = .data$label,
+        names_from = {{ columns_var }},
         values_from = .data$result
         )
   }
@@ -38,9 +43,15 @@ pivot_freqs <-
 
 pivot_errors <- function(dataset) {
 
-  if (unique(dataset$label)[1] == "" &
+  if (unique(dataset$label)[1] == '' &
       length(unique(dataset$label)) == 1) {
-    stop("Your frequencies label column is blank. Please provide labels on which to pivot")
+    stop('Your frequencies label column is blank. Please provide labels on which to pivot')
+  }
+
+  if (
+    sum(names(dataset) %in% 'group_var') != 1
+  ) {
+    stop('Your frequencies does not contain a group_var. It must have a group_var to pivot correctly')
   }
 
 }